Seperate tests to different files
